+++ service/controller/release/resource/status/create.go
@@ -54,7 +54,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	}
 
 	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("setting status for release %#q in namespace %#q", release.Name, release.Namespace))
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("setting status for release %#q", release.Name))
 
 		release.Status.Ready = releaseDeployed
 		err := r.k8sClient.CtrlClient().Status().Update(
@@ -65,7 +65,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 			return microerror.Mask(err)
 		}
 
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("status set for release %#q in namespace %#q", release.Name, release.Namespace))
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("status set for release %#q", release.Name))
 	}
 
 	return nil