+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add monitoring labels and add basic labels
+
 ## [2.0.0] - 2020-07-23
 
 ### Added