+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add monitoring labels and add basic labels
 
+## Changed
+
+- Updated backward incompatible Kubernetes dependencies to v1.18.5.
+
 ## [2.0.0] - 2020-07-23
 
 ### Added