+++ CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add monitoring labels and add basic labels
 
+## Changed
+
+- Updated backward incompatible Kubernetes dependencies to v1.18.5.
+- Don't error when app was not found while deleting app.
+- Deleted Release namespace from logging since releases are not namespaced.
+
 ## [2.0.0] - 2020-07-23
 
 ### Added